requirements.txt
click>=7.1.2
Flask>=1.1.2
itsdangerous>=1.1.0
Jinja2>=2.11.2
lxml>=4.5.2
MarkupSafe>=1.1.1
pycrypto>=2.6.1
Werkzeug>=1.0.1
